requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azapi==3.0.8
beautifulsoup4==4.11.2
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
install==1.3.5
multidict==6.0.4
mutagen==1.46.0
nextcord==2.3.3
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.4.0
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
typing_extensions==4.5.0
urllib3==1.26.14
websockets==10.4
yarl==1.8.2
yt-dlp==2023.2.17